create main.cpp and student class
Creating a student class equiped with the constructors which will allows us to create a data structure with different information by students which will later on be used for sorting
those student.
there is also a display function to display the student information into the terminal.
main.cpp isn't working at the moment because i will create a manager to add, sub and update students first